melk: Don't clobber output files on error
Revamp melk to using temporary files for output, and rename them in
place to overwrite old files only if all processing was successful.

This avoids an error creating a 0-byte output file, and a subsequent
make picking that up as the data to be used resulting in empty
files being published.